refactor: remove unused imports and simplify import
