+++ src/containers/SidePanel/Header/Header.tsx
@@ -3,7 +3,6 @@ import html2canvas from 'html2canvas';
 
 import { Block } from 'baseui/block';
 import { useDispatch, useSelector } from 'react-redux';
-import { AnyAction, Dispatch } from 'redux';
 import { GraphList } from '../../Graph';
 import { getGraphList, getUI, setName } from '../../../redux';
 import * as Icon from '../../../components/Icons';
